pyproject.toml
[project]
name = "expander"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Major Hayden", email = "major@mhtx.net" }]
requires-python = ">=3.13,<3.14"
dependencies = [
    "docling",
    "docling-core",
    "docling-parse",
    "httpx>=0.28.1",
    "pgvector>=0.4.1",
    "psycopg>=3.2.9",
    "pydantic-settings>=2.10.1",
    "rich>=14.1.0",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]

[project.scripts]
expander = "expander:main"

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-httpx>=0.35.0",
    "pytest-randomly>=3.16.0",
    "python-gitlab>=6.1.0",
    "ruff>=0.12.4",
    "tox>=4.28.1",
]

[tool.pytest.ini_options]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-branch",
    "--color=yes",
]
# Having problems with tests? Uncomment the following line to see more output.
# log_cli = false
# log_level = "DEBUG"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = ["tests"]

[tool.pyright]
venvPath = "."
venv = ".venv"
